+++ scilab/modules/ast/includes/types/graphichandle.hxx
@@ -82,6 +82,8 @@ public :
 
     virtual bool invoke(typed_list & in, optional_list & opt, int _iRetCount, typed_list & out, const ast::Exp & e) override ;
 
+    bool getMemory(int* _piSize, int* _piSizePlusType);
+
 protected :
     inline ScilabType		getType(void)
     {